created update product route

+++ Backend_e-comm/controller/product.controller.js
@@ -4,6 +4,9 @@ const { validationResult } = require("express-validator");
 const fs = require("fs");
 //! Adding product
 const addProduct = async (req, res) => {
+  if (!req.file) {
+    return res.status(400).json({ error: "Image must be required." });
+  }
   const errors = validationResult(req);
   if (!errors.isEmpty()) {
     fs.unlink(req.file.path, (err) => {
@@ -12,12 +15,18 @@ const addProduct = async (req, res) => {
     return res.status(400).json({ errors: errors.array() }); // ! bad request
   } else {
     try {
-      const {category_name,product_name,product_url_slug,description,price,product_status,stock_quantity} = req.body;
+      const {
+        category_name,
+        product_name,
+        description,
+        price,
+        product_status,
+        stock_quantity,
+      } = req.body;
       const checkCategory = await Category.findOne({ category_name });
       // ! if category don't exists send invalid category -> else get category id
       if (!checkCategory) {
         throw new Error("Invalid category");
-        return res.status(400).json({ error: "Invalid category." });
       }
       let category_id = await Category.findOne({ category_name }, { _id: 1 });
       category_id = category_id._id;
@@ -28,7 +37,7 @@ const addProduct = async (req, res) => {
       const Product_info = {
         product_name,
         image: dbfilepath,
-        product_url_slug,
+        product_url_slug: product_name.replaceAll(" ", "-"),
         category_id,
         description,
         price,
@@ -55,21 +64,30 @@ const addProduct = async (req, res) => {
       fs.unlink(req.file.path, (err) => {
         err && res.send(err);
       });
+
       res.status(500).json({ success: false, error: error.message });
     }
   }
 };
+//! updating image gallery
 const updateProduct = async (req, res) => {
   const errors = validationResult(req);
   if (!errors.isEmpty()) {
-    return res.status(400).json({ errors: errors.array() }); // 400 means bad request
+    if (req.files) {
+      req.files.map((file) => {
+        fs.unlink(file.path, (err) => {
+          err && res.send(err); // :console.log("files deleted successfully.")
+        });
+      });
+    }
+    return res.status(400).json({ errors: errors.array() }); // ! bad request
   } else {
     try {
       const product_id = req.params.id;
+      !product_id && new Error("product id not found.");
       const {
         category_name,
         product_name,
-        product_url_slug,
         description,
         price,
         product_status,
@@ -81,21 +99,40 @@ const updateProduct = async (req, res) => {
       }
       category_id = category_id._id;
       let product_images = [];
-      req.files.map((file) => {
+      req.files.forEach((file) => {
         product_images.push(
-          "/" + file.path.split("/")[1] + "/" + file.path.split("/")[2]
+          `/${file.path.split("/")[1]}/${file.path.split("/")[2]}`
         );
       });
-      const updateProduct_info = {
-        product_name,
-        product_images,
-        product_url_slug,
-        category_id,
-        description,
-        price,
-        stock_quantity,
-        product_status,
-      };
+      const fileslength = req.files.length;
+      let updateProduct_info;
+      if (fileslength > 0) {
+        updateProduct_info = {
+          product_name,
+          product_images,
+          product_url_slug: product_name.replaceAll(" ", "-"),
+          category_id,
+          description,
+          price,
+          stock_quantity,
+          product_status,
+        };
+      } else {
+        updateProduct_info = {
+          product_name,
+          product_url_slug: product_name.replaceAll(" ", "-"),
+          category_id,
+          description,
+          price,
+          stock_quantity,
+          product_status,
+        };
+      }
+      //? if product images array length is 4 throw Error
+      const product_imagesArray = await Product.findById(product_id,{product_images:1})
+      if(product_imagesArray.product_images.length===4){
+        throw new Error('Maximum 4 images can be uploaded.')
+      }
       const updateproduct = await Product.findByIdAndUpdate(product_id, {
         $set: updateProduct_info,
       });
@@ -109,10 +146,18 @@ const updateProduct = async (req, res) => {
             .status(400)
             .json({ success: false, error: "product does not exist." });
     } catch (error) {
-      error.message === "" &&
+      if (error.message === "") {
         res
           .status(500)
           .json({ success: false, error: "Internal server error." });
+      }
+      if (req.files) {
+        req.files.map((file) => {
+          fs.unlink(file.path, (err) => {
+            err && res.send(err); // :console.log("files deleted successfully.")
+          });
+        });
+      }
       res.status(500).json({ success: false, error: error.message });
     }
   }
